requirements.txt
# 最小構成のビルドに必要
setuptools>=58.0.0
wheel

# Celery本体とRedisブローカー
celery==5.3.6
redis==5.0.1

# 環境変数の読み込み
python-dotenv==1.0.0
requests==2.31.0
